requirements.txt
python-telegram-bot==21.0
requests>=2.31.0
psutil>=5.9.0
matplotlib>=3.5.0
numpy>=1.21.0
PyPDF2>=3.0.0
python-docx>=1.0.0
Pillow>=10.0.0
